tests(tracing): move test mocks to test_helpers_test.go and restore helpers
