+++ platform/view/services/tracing/provider_test.go
@@ -12,8 +12,10 @@ import (
 
 	"github.com/stretchr/testify/require"
 	"go.opentelemetry.io/otel/trace"
+	"go.opentelemetry.io/otel/trace/noop"
 
 	"github.com/hyperledger-labs/fabric-smart-client/platform/view/services/metrics"
+	"github.com/hyperledger-labs/fabric-smart-client/platform/view/services/metrics/disabled"
 	"github.com/hyperledger-labs/fabric-smart-client/platform/view/services/tracing"
 )
 
@@ -81,49 +83,6 @@ func TestTracerProvider_TracerWithMetricsOpts(t *testing.T) {
 	require.NotNil(t, span)
 }
 
-type mockMetricsProvider struct {
-	counters   map[string]*mockCounter
-	histograms map[string]*mockHistogram
-}
-
-type mockCounter struct {
-	values map[string]float64
-}
-
-type mockHistogram struct {
-	values map[string][]float64
-}
-
-func (c *mockCounter) Add(delta float64) {
-	c.values["total"] += delta
-}
-
-func (c *mockCounter) With(_ ...string) metrics.Counter {
-	return c
-}
-
-func (h *mockHistogram) Observe(value float64) {
-	h.values["observations"] = append(h.values["observations"], value)
-}
-
-func (h *mockHistogram) With(_ ...string) metrics.Histogram {
-	return h
-}
-
-func (mp *mockMetricsProvider) NewCounter(opts metrics.CounterOpts) metrics.Counter {
-	key := fmt.Sprintf("%s_%s_%s", opts.Namespace, opts.Subsystem, opts.Name)
-	c := &mockCounter{values: make(map[string]float64)}
-	mp.counters[key] = c
-	return c
-}
-
-func (mp *mockMetricsProvider) NewHistogram(opts metrics.HistogramOpts) metrics.Histogram {
-	key := fmt.Sprintf("%s_%s_%s", opts.Namespace, opts.Subsystem, opts.Name)
-	h := &mockHistogram{values: make(map[string][]float64)}
-	mp.histograms[key] = h
-	return h
-}
-
 func TestTracerProvider_CompleteTraceLifecycle(t *testing.T) {
 	t.Parallel()
 
@@ -163,6 +122,29 @@ func TestGetProvider_Success(t *testing.T) {
 	retrievedProvider, err := tracing.GetProvider(mockServiceProvider{provider: provider})
 	require.NoError(t, err)
 	require.Equal(t, provider, retrievedProvider)
+}
+
+func newTestProvider(t *testing.T) tracing.Provider {
+	t.Helper()
+
+	provider := tracing.NewProviderWithBackingProvider(noop.NewTracerProvider(), &disabled.Provider{})
+	require.NotNil(t, provider)
+
+	return provider
+}
+
+func newTestTracer(t *testing.T, labelNames ...string) trace.Tracer {
+	t.Helper()
+
+	provider := newTestProvider(t)
+	tracer := provider.Tracer("test-tracer", tracing.WithMetricsOpts(tracing.MetricsOpts{
+		Namespace:  "test",
+		Subsystem:  "test",
+		LabelNames: labelNames,
+	}))
+	require.NotNil(t, tracer)
+
+	return tracer
 }
 
 type mockServiceProvider struct {
