+++ platform/view/services/tracing/test_helpers_test.go
@@ -7,35 +7,50 @@ SPDX-License-Identifier: Apache-2.0
 package tracing_test
 
 import (
-	"testing"
+	"fmt"
 
-	"github.com/stretchr/testify/require"
-	"go.opentelemetry.io/otel/trace"
-	"go.opentelemetry.io/otel/trace/noop"
-
-	"github.com/hyperledger-labs/fabric-smart-client/platform/view/services/metrics/disabled"
-	"github.com/hyperledger-labs/fabric-smart-client/platform/view/services/tracing"
+	"github.com/hyperledger-labs/fabric-smart-client/platform/view/services/metrics"
 )
 
-func newTestProvider(t *testing.T) tracing.Provider {
-	t.Helper()
-
-	provider := tracing.NewProviderWithBackingProvider(noop.NewTracerProvider(), &disabled.Provider{})
-	require.NotNil(t, provider)
-
-	return provider
+type mockMetricsProvider struct {
+	counters   map[string]*mockCounter
+	histograms map[string]*mockHistogram
 }
 
-func newTestTracer(t *testing.T, labelNames ...string) trace.Tracer {
-	t.Helper()
+type mockCounter struct {
+	values map[string]float64
+}
 
-	provider := newTestProvider(t)
-	tracer := provider.Tracer("test-tracer", tracing.WithMetricsOpts(tracing.MetricsOpts{
-		Namespace:  "test",
-		Subsystem:  "test",
-		LabelNames: labelNames,
-	}))
-	require.NotNil(t, tracer)
+type mockHistogram struct {
+	values map[string][]float64
+}
 
-	return tracer
+func (c *mockCounter) Add(delta float64) {
+	c.values["total"] += delta
+}
+
+func (c *mockCounter) With(_ ...string) metrics.Counter {
+	return c
+}
+
+func (h *mockHistogram) Observe(value float64) {
+	h.values["observations"] = append(h.values["observations"], value)
+}
+
+func (h *mockHistogram) With(_ ...string) metrics.Histogram {
+	return h
+}
+
+func (mp *mockMetricsProvider) NewCounter(opts metrics.CounterOpts) metrics.Counter {
+	key := fmt.Sprintf("%s_%s_%s", opts.Namespace, opts.Subsystem, opts.Name)
+	c := &mockCounter{values: make(map[string]float64)}
+	mp.counters[key] = c
+	return c
+}
+
+func (mp *mockMetricsProvider) NewHistogram(opts metrics.HistogramOpts) metrics.Histogram {
+	key := fmt.Sprintf("%s_%s_%s", opts.Namespace, opts.Subsystem, opts.Name)
+	h := &mockHistogram{values: make(map[string][]float64)}
+	mp.histograms[key] = h
+	return h
 }
